Okay, next thing was browser visualization.
Separated the core and browser areas. Gulpfile is now ridiculous, but it'll
get fixed later.

Threw together a little react-based visualization that renders an arbitrary
grid / solution / calculated regions. It's hideous, and clunky, but it works
for testing purposes for now.

Super impressed with browserify, So far, no real need to mess around with
bower.