Beginnings of individual puzzle rules
Category seems to be working all right, other than it showing me that naive
backtracking may be a non-starter. Can't seem to prune the search tree early
enough; a couple category nodes near the beginning, and a 9x9 is intractable.

But, did get the React stuff hooked up to the solver, and displaying the
retrieved solutions. So that's something.

And for smaller stuff, we can now solve the color puzzles. Rule order is
probably going hexagons -> stars -> triangles -> tetris.